refactor(optimizations): extract shared useNavigateToOptimizationStudio hook (OPIK-6761)
Address review feedback: OptimizationsEmptyState.navigateToStudio and
StudioTemplates.handleTemplateClick duplicated the same project-scoped
/optimizations/new routing. Extract a shared useNavigateToOptimizationStudio
hook and use it in both, so the routing lives in one place. No behavior change.
